+++ domain/src/main/java/com/ieum/domain/model/address/Address.kt
@@ -1,0 +1,7 @@
+package com.ieum.domain.model.address
+
+data class Address(
+    val code: String,
+    val name: String,
+    val fullName: String,
+)
