--- domain/src/main/java/com/ieum/domain/model/common/Address.kt
@@ -1,6 +1,0 @@
-package com.ieum.domain.model.common
-
-data class Address(
-    val key: String,
-    val name: String,
-)
