--- domain/src/main/java/com/ieum/domain/model/common/SGISToken.kt
@@ -1,5 +1,0 @@
-package com.ieum.domain.model.common
-
-data class SGISToken(
-    val accessToken: String,
-)
