+++ domain/src/main/java/com/ieum/domain/repository/AddressRepository.kt
@@ -1,0 +1,9 @@
+package com.ieum.domain.repository
+
+import com.ieum.domain.model.address.Address
+
+interface AddressRepository {
+    suspend fun getAccessToken(): String
+
+    suspend fun getAddressList(accessToken: String, code: String): List<Address>
+}
