+++ domain/src/main/java/com/ieum/domain/usecase/address/GetSGISTokenUseCase.kt
@@ -1,0 +1,12 @@
+package com.ieum.domain.usecase.address
+
+import com.ieum.domain.repository.AddressRepository
+import javax.inject.Inject
+
+class GetSGISTokenUseCase @Inject constructor(
+    private val addressRepository: AddressRepository,
+) {
+    suspend operator fun invoke(): Result<String> = runCatching {
+        addressRepository.getAccessToken()
+    }
+}
