--- domain/src/main/java/com/ieum/domain/usecase/common/GetSGISTokenUseCase.kt
@@ -1,12 +1,0 @@
-package com.ieum.domain.usecase.common
-
-import com.ieum.domain.repository.SGISRepository
-import javax.inject.Inject
-
-class GetSGISTokenUseCase @Inject constructor(
-    private val sgisRepository: SGISRepository,
-) {
-    suspend operator fun invoke() = runCatching {
-        sgisRepository.getAccessToken()
-    }
-}
